Move helpers into smaller modules Integrate storage and parser Make lang parser a pkg
